+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
  * removed `BindingState::UNDECIDED`
  * removed `InstallInfo::*Enabled*()` methods
  * added config key "bootstrap-file" which is loaded before loading plugins
+ * added repository type "path-mapping"
 
 * 1.0.0-beta5 (2015-05-29)